+++ app/components/board/post/index.ts
@@ -10,7 +10,6 @@ import NewsfeedService from 'potber-client/services/newsfeed';
 import Thread from 'potber-client/models/thread';
 import SettingsService, { AvatarStyle } from 'potber-client/services/settings';
 import RendererService from 'potber-client/services/renderer';
-import { scrollToHash } from 'ember-url-hash-polyfill';
 import LocalStorageService from 'potber-client/services/local-storage';
 
 interface Signature {
@@ -18,7 +17,6 @@ interface Signature {
     post: Post;
     thread: Thread;
     subtle?: boolean;
-    isFinalElement?: boolean;
   };
 }
 
@@ -37,6 +35,10 @@ export default class PostComponent extends Component<Signature> {
   }
 
   declare args: Signature['Args'];
+
+  get elementId() {
+    return `post-${this.args.post.id}`;
+  }
 
   get date() {
     return new Date(this.args.post.date).toLocaleString();
@@ -142,33 +144,6 @@ export default class PostComponent extends Component<Signature> {
       return `${this.args.post.editedCount}x bearbeitet, zuletzt von ${
         this.args.post.lastEdit.user.name
       } am ${new Date(this.args.post.lastEdit.date).toLocaleString()}`;
-    }
-  }
-
-  @action updateScrollPosition() {
-    // Check whether the URL contains a post id that matches this post
-    const params = new URL(window.location.href).searchParams;
-    if (params.has('PID') && params.get('PID')) {
-      // If post id was supplied, we also need to add the anchor
-      // Set the hash without triggering without triggering a browser scroll action
-      const currentState = { ...history.state };
-      history.replaceState(
-        currentState,
-        'unused',
-        `#reply_${params.get('PID')}`
-      );
-      if (window.location.hash) {
-        scrollToHash(`reply_${params.get('PID')}`);
-      }
-    } else if (params.get('scrollToBottom') === 'true') {
-      // if scrollToBottom was supplied, scroll to bottom
-      this.renderer.trySetScrollPosition({
-        top: document.body.scrollHeight,
-        behavior: 'smooth',
-      });
-    } else if (this.args.isFinalElement) {
-      // Else, we will reset the scroll position after the last post has been rendered
-      this.renderer.trySetScrollPosition();
     }
   }
 }